study_pr/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.6
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.6.0
django-filter==22.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
kombu==5.2.4
Markdown==3.4.1
Pillow==9.2.0
prompt-toolkit==3.0.38
python-dotenv==0.21.0
pytz==2022.5
redis==4.5.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.6
